+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 0.0.1-alpha.2 (2021-03-06) 
+
+## Added Features
+
+ - iOS Reading Base64 images
+ - iOS Sharing an image
+ - iOS Gestures for Zooming in/out
+ - maxzoomscale options
+ - compression quality options
+
+
 # 0.0.1-alpha.1 (2021-03-02) 
 
 - Initial commit for Photoviewer Android and iOS